Apply as of date event filtering.
This commit allows a user to pass an as of date into the client overrides configuration file. The application will then recognize this date as the 'run date' and any leads created after this date will be filtered from scoring. In addition, if a date is supplied then this date will be used as the lookback date as well.